Server2/requirements.txt
flask
pymysql
mysql-connector-python
cryptography